fix: -Werror=conversion in ufsecp_musig2.cpp + API drift in shim_security_v7 test
src/cpu/src/impl/ufsecp_musig2.cpp:288 (ufsecp_musig2_partial_sign_v2):
  CT byte-compare loop `diff |= (a[i] ^ b[i])` promotes uint8_t through int.
  Under -Werror=conversion this becomes a build error. Cast back to uint8_t
  explicitly. No behavior change — the bitwise OR width is unchanged.

  Caught by GitHub `Security Audit / Build with -Werror`. Root cause of the
  cascading failures across Sanitizers (ASan+UBSan / MSan / TSan), Valgrind,
  CI / linux (clang+gcc Debug+Release), linux-arm64/riscv64, macos, wasm,
  Benchmark Dashboard, CAAS Security Gates — all of these share the same
  build configuration, so a single -Werror failure blocked all jobs.

audit/test_regression_shim_security_v7.cpp:
  - secp256k1_musig_pubkey_agg(ctx, &agg_pk, &keyagg, pubs, n) — drop the
    extra `nullptr` first slot. The old prototype had a 6-arg form with
    a scratch-space pointer; the current shim API matches upstream's 5-arg
    `(ctx, agg_pk, keyagg_cache, pubkeys, n)`.
  - secp256k1_musig_nonce_gen(ctx, secnonce, pubnonce, session_id32, seckey,
    pubkey, msg32, keyagg_cache, extra_input32) — pass `sk1` (raw 32-byte
    seckey) and `&pub1` (secp256k1_pubkey*) instead of `&kp1` (keypair*).
    The old prototype accepted a keypair; the current shim API takes the
    seckey + pubkey pair separately, matching upstream MuSig2 nonce-gen
    requirements (BIP-327 §NonceGen).

  Caught by GitHub `Benchmark Smoke Check` build phase.

Local verification:
  - cmake -DSECP256K1_WERROR=ON build of ufsecp_static → succeeds.
  - cmake --build out/cpu-release --target unified_audit_runner → succeeds.
  - ci_local.sh quick mode: ALL 19 CHECKS PASSED (2 advisory-skipped).

+++ src/cpu/src/impl/ufsecp_musig2.cpp
@@ -283,9 +283,11 @@ ufsecp_error_t ufsecp_musig2_partial_sign_v2(
 
     const uint8_t* expected_pubkey33 = pubkeys + signer_index * 33;
     // Constant-time comparison to avoid branch-based timing leak on secret key.
+    // Use unsigned widening to silence -Wconversion: uint8_t ^ uint8_t in C++
+    // promotes through int, so the result must be explicitly narrowed back.
     uint8_t diff = 0;
     for (size_t i = 0; i < 33; ++i) {
-        diff |= (derived_compressed[i] ^ expected_pubkey33[i]);
+        diff |= static_cast<uint8_t>(derived_compressed[i] ^ expected_pubkey33[i]);
     }
     if (SECP256K1_UNLIKELY(diff != 0)) {
         secp256k1::detail::secure_erase(&sk, sizeof(sk));
